Fix in l3g4200d

+++ src/main/drivers/accgyro_l3g4200d.c
@@ -54,7 +54,7 @@
 #define L3G4200D_DLPF_78HZ       0x80
 #define L3G4200D_DLPF_93HZ       0xC0
 
-static void l3g4200dInit(uint16_t lpf);
+static void l3g4200dInit(uint8_t lpf);
 static bool l3g4200dRead(int16_t *gyroADC);
 
 bool l3g4200dDetect(gyro_t *gyro)
@@ -76,7 +76,7 @@ bool l3g4200dDetect(gyro_t *gyro)
     return true;
 }
 
-static void l3g4200dInit(uint16_t lpf)
+static void l3g4200dInit(uint8_t lpf)
 {
     bool ack;
